requirements.txt
pyTelegramBotAPI
python-dotenv
yt-dlp
